Introduce retries to `databricks psql` command  (#3492)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

Sometimes the connection to the database instance might fail, which can
be remediated with a retry. This change makes the command to retry the
connection 3 times by default.

To disable the retries users can set a `max-retries` value to zero:

```
$ databricks psql my-instance --max-retries 0
```

## Tests
<!-- How have you tested the changes? -->
Added a new acceptance test, modified and existing acceptance test; ran
manual tests on Mac and Windows

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->