Implement database_catalog in direct deployment (#3488)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This change allows to deploy database catalogs when using `direct-exp`
deployment method

## Tests
<!-- How have you tested the changes? -->
Existing test is expanded to run using `direct-exp`

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->