dyn: Allow interpolation for non-string primitive (#3522)
## Why
Terraform supports that, so to match $resources resolution we need to
support it as well. Also, it's natural for interpolation to support
types that have string representation.

## Tests
Existing unit test, new acceptance test for $var case. In my remote
state PR, I have acc test for $resources case as well.