Add "ssh" and "ssh setup" commands (#3470)
## Changes

This is the first PR that introduces the ssh-tunnel functionality into
the CLI

See README for an overview of the SSH tunnel feature and its usage. 

This PR only adds `ssh setup` command, `connect` and `server` are here:
- https://github.com/databricks/cli/pull/3471
- https://github.com/databricks/cli/pull/3475

The whole `ssh` command group is hidden (not available in `--help`
output, and in this PR it's not even added to the root cmd), and the
usage text has private preview disclaimer. Hidden flag will be removed
when we move to public preview phase (no timeline yet).

`setup` command updates local ssh config (`~/.ssh/config`). The config
entry relies on `ProxyCommand` option calling `databricks ssh connect`,
which is implemented in a follow up PR.


## Why
We've agreed to move the implementation form the separate repo to the
main databricks CLI

## Tests
Manually and unit tests.

---------

Co-authored-by: Pieter Noordhuis <pieter.noordhuis@databricks.com>