Moved setting default source for tasks to tfdyn (#3528)
## Changes
Moved setting default source for tasks to tfdyn

## Why
This prevents defaults from being modified after PyDABs have access to
them.
This fix is "backend" fixup so it shouldn't be in resource mutators
anyway

## Tests
Existing acceptance tests pass.

It required splitting the output for TF and direct because in direct
mode, we don't need to send the source field unless it's explicitly set