Add top level run_as support for DLT pipelines (#3307)
## Changes
This PR adds support for top-level run_as to DABs. Before DABs would
error if pipelines were used with the `run_as` field set to a value.

Now, we'll transparently read the run_as value and set it for all
pipelines in the bundle.

## Tests
New acceptance tests. Also manually tested that the run_as feature works
as expected.