Allow referencing job libraries outside bundle root without the need to specify sync root (#2842)
## Changes
Allow referencing job libraries outside bundle root without the need to
specify sync root.

Previously it was failing with error indicating the libraries are not
sync root, now it does not fail anymore.

## Why
Libraries has their own upload cycle and do not rely on general bundle
sync mechanism therefore it is not necessary to specify or check that
libraries are in sync root.

## Tests
Added acceptance tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->