acc: sort output in debuglog (#3592)
Fix order issues in the output by sorting the output.

Previous attempt to fix this test did not work:
https://github.com/databricks/cli/pull/3590

The flakiness started after rate limit was disabled in
https://github.com/databricks/cli/pull/3581

For context, this test was originally sorting the output but that was
disabled in https://github.com/databricks/cli/pull/2414 because it
seemed that output was stable. It was only stable because of delays
introduced by rate limiter.