Add cluster selector UI and auto-start option to ssh commands (#3607)
## Changes
- Add a cluster selector UI to the `ssh setup` command. We don't filter
or check the clusters during that stage, as it's done later by the
`validateClusterAccess`. Cluster option is no longer required for the
`setup` command. A few customers asked for this already.
- Add an `--auto-start-cluster` flag to both `ssh setup` and `ssh
connect` commands, which defaults to true. Useful for PyCharm users,
where the IDE checks every host in the config to see the if the
connection can be established, starting all the clusters.
- Change a few `client.go` functions to accept ClientOptions instead of
a long list of separate arguments


Based on https://github.com/databricks/cli/pull/3569

## Tests
No unit tests, integration tests will be in the follow ups.