Fix `checkForPreventDestroy` to check all resources if some does not have prevent_destroy set (#3615)
## Changes
Previously, `checkForPreventDestroy` aborted earlier if it encountered a
resource with no prevent_destroy set, and wouldn't check the rest of the
resources.
Now this is fixed to check all resources.

## Why
Fixes a bug in `checkForPreventDestroy`

## Tests
Added a regression unit test for `checkForPreventDestroy`

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->