Skip fields generation for resource_quality_monitor in tf codegen (#3462)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

When running codegen, a developer who is not aware of the issue with
fields of the quality monitor can get confused / spend a lot of time
trying to understand the issue with it. Skipping this generation helps
to avoid this confusion. The downside of this change is that new changes
to the quality monitor will not be automatically picked up by the
codegen. To mitigate that a warning line in the log is added, so that if
you are looking for changes specifically in the quality monitor, you
will get alerted by the message.