Introduce new bundle variable: `${workspace.current_user.domain_friendly_name}` (#3623)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

`${workspace.current_user.dns_name}` will allow bundle authors to prefix
apps and database instances resources with caller's usernames (useful
for development targets). Currently `short_name` does not always fit
this purpose because it can contain characters that apps / database
instances do not allow in the names of resources.

## Tests
<!-- How have you tested the changes? -->
- Existing tests
- New unit test with transformation examples
- Manual check that `dns_name` can be used as a bundle variable

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->