Fixed bundle deploy to not update permissions for unbound resources (#3642)
## Changes
Fixed bundle deploy to not update permissions for unbound resources

## Why
The original issue occurred because we hadn't removed the permissions
section for the corresponding resources from the TF state. and
therefore, permissions were continued to be managed by TF and, as a
result, cleared out.

## Tests
Added an acceptance test

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->