Include skip actions in JSON plan (#3663)
## Changes
Include 'skip' nodes in JSON plan.

## Why
We need to process 'skip' resources as well (e.g. to resolve
references). Today, we have these nodes in the graph but not in the
plan. Once we move to serialized plan
https://github.com/databricks/cli/pull/3636 we will build graph from the
plan so they must match. This prepares for that (and minimizes that PR).

This also gives more information to users of JSON plan, e.g. they can
count unchanged count.

## Tests
Existing tests.