internal: update GetResourceConfig to use dyn/convert rather than JSON roundtrip (#3646)
## Why
For direct, we fetch configs as struct instances via GetResourceConfig,
which is using json roundtrip between dyn value and struct. That,
however, does not work well when there references in the config, see
https://github.com/databricks/cli/pull/3645 for details.

This switches to converting dyn.Value with dyn/convert ToTyped function.
The function was not used previously because it did not handle
ForceSendFields correctly for embedded structs which caused JSON
marshaller in Go SDK to return error. This was fixed in
https://github.com/databricks/cli/pull/3650 and
https://github.com/databricks/cli/pull/3649

## Tests
Existing tests.