direct: Detailed bundle plan (#3636)
## Changes

The JSON plan (output of “bundle debug plan”) is now input to “bundle
deploy” internally. This means it contains full information about
resources about to be deployed.

The JSON plan stores configs as well as all unresolved fields as map
(path -> reference). There is new module, structvar, that supports
resolving these references. It does it quickly and without copying the
full config for every resolution, like the previous implementation did.

The JSON plan also stores per-field action. This is purely
informational, it gives an idea why a certain action for the resource
was chosen.

During plan and apply, variable resolution is now done in the node that
contains the reference. Previously, it was updated by the node that was
referenced. This results in better error messages when resolution does
not work and it does not fail resources that themselves have no issues
(but have incorrect references pointing to them).

References can now refer to missing field and it works like in terraform
(missing_string_field test).

There is slight simplification in how $resources are resolved. All local
resolutions (from config) are now done in plan phase. All remote
resolutions (from remote state) are now done in apply phase. Previously
we could resolve some remote references in plan stage if action was
“skip”. This might result in slower deploys, so this is a potential perf
regression, to be revisited in the future.

Depends on https://github.com/databricks/cli/pull/3650 &
https://github.com/databricks/cli/pull/3646

Other changes:
- dagrun.Graph now exports its fields. dagrun.OutgoingLabels() is
removed.
- new function structaccess.Set() to set any field in a struct by path
(used by structvar).
- dresources.ClassifyByTriggers([]Change) is replaced by
dresources.ClassifyByTriggers(Change), aggregation logic now lives in
bundle_plan.go

## Why
Having the plan as input to deploy provides a view into deployment that
we did not have before, allowing to see exactly what is going to be
deployed and why.

This is useful for:
 - debugging (why is my pipeline being recreated?)
 - see diff between deployed infra and config
- deploying exactly what was in the diff and nothing else: "bundle
deploy -readplan plan.json" can be implemented easily after this change.