direct: Added support for ClassifyChanges (#3690)
## Changes
Added support for ClassifyChanges

## Why
Needed for https://github.com/databricks/cli/pull/3682

Some resources might have different types of changes based on their
remote state. For example, the cluster's resize action becomes an update
if the cluster is stopped. In order to do such classification, we need
access to the remote state, so we are introducing a new ClassifyChange
method.

ClassifyChange is called first, if it returns ActionTypeUnset,
ClassifyByTriggers called

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->