Remove check for interactive mode in labs installer (#3709)
## Changes

If this condition evaluates to true, it returns `(nil, nil)` and leads
to a panic further along in the code. The check for an interactive
terminal is done earlier than necessary and can be removed altogether.
The `askWorkspaceProfile` function also checks for an interactive
terminal and returns an error if there isn't one.

We mark Git Bash as a terminal where prompting is not supported (see
#1069).

Fixes #3597.

## Tests

Manually confirmed the installer no longer panics in Git Bash on
Windows.