Remove empty permissions from json output (#3715)
## Changes
When processing permissions, do not convert nils to empty slice in dyn
value.

## Why
Without this '"permissions: []"' always shows up in JSON output. When
permissions are not present in the config, this is confusing.

In addition, Permissions fields are marked as "omitempty" so even if
they are present, they should be cleaned up.

Related:
https://github.com/databricks/cli/pull/2756#pullrequestreview-2790052042

## Tests
Existing tests.