Move tools to tools/go.mod; add golangci-lint there; update it to 2.5.0 (#3722)
## Changes
- Move yamlfmt, gotestsum, and golangci-lint into a tools module that
pins their versions
- This pins golangci-lint as 2.5.0 which adds "no naked return" in
gofumpt so this PR updates the code that triggers it.
- We no longer use golangci-lint github action, the same command/version
that is configured in tools/go.mod is used on CI as well.
- Pre-build golangci-lint in Makefile, this takes 0.5s of start time on
my machine.
- Configure dependabot to manage tools/go.mod

## Why
Relying on system golangci-lint results in people (and agents) running
into issue due to version mismatch. This manages the version inside the
repo, so everyone is one the same version and updates go through PR +
testing.

------
https://chatgpt.com/codex/tasks/task_e_68e1967083908325baa2939b043cae54

Original PR: https://github.com/denik/cli/pull/3