Respect `BROWSER` environment variable in "auth login" command (#3659)
## Changes

Respect the `BROWSER` environment variable for the OAuth login flow.
When set, the value is used as executable to launch the URL to
authenticate the user. When set to "none", the CLI prints a message with
the URL to open. When not set, it defers to the existing logic to open a
browser via the `pkg/browser` package.

## Why

Context in:
* https://github.com/databricks/cli/issues/3464
* https://github.com/databricks/databricks-vscode/issues/1764

Also refer to issues linked in the description of these issues for
broader context, and why `pkg/browser` doesn't respect `BROWSER` itself.
It boils down to it not being a ratified standard.

We choose to respect it to make the interactive login flow from VS Code
running in a dev container work.

## Tests

New acceptance test that goes through the login flow via the test
server.