Add a flag to configure DB Connect Serverless profiles to `databricks auth login`. (#3728)
## Changes

This PR extends `databricks auth login` with a new
`--configure-serverless` flag that triggers the flow to configure the
profile for DB Connect Serverless.

The serverless configuration flow simply takes care of configuring the
profile by setting `serverless_compute_id = auto` and deleting the value
of `cluster_id` if any. Though, it is meant to be extended to prompt
users to provide additional properties.

Side note: this PR illustrates the opportunity to consolidate the
profile code management.

## Tests

New acceptance test.

---------

Co-authored-by: Pieter Noordhuis <pieter.noordhuis@databricks.com>