acc: Separate test output for direct and terraform for interactive_single_user test (#3768)
## Changes
Separate test output for direct and terraform for
interactive_single_user test

## Why
TF test behaviour for wheel jobs on single-user interactive clusters is
technically incorrect because TF in such a setup produces permanent
drift (incorrect), which leads to cluster restart on each bundle deploy,
and therefore wheel changes are picked up
```
  # databricks_cluster.shared_autoscaling will be updated in-place
  ~ resource "databricks_cluster" "shared_autoscaling" {
        id                             = "1015-102652-xxxxxxx"
      - single_user_name               = "aaaaaaaabbbb-bbbb-cccc-dddd-bb6ce382668a" -> null
        # (19 unchanged attributes hidden)

        # (1 unchanged block hidden)
    }
```
Direct does not update/restart the cluster on subsequent deploys
(correct), and changes for the wheel are not picked, which is expected.

I separated the test output into 2 separate files for now, to unlock the
test.

## Tests
Existing tests pass

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->