Remove unused JSON progress logging mode (#3812)
## Changes

The progress logger now only supports `append` mode.

## Why

JSON mode was added in #276 (March 2023). The intent was to make
progress events machine readable but this didn't materialize (we didn't
end up using it in the VS Code extension).

This functionality would print events to stderr if the user specified
`--progress-format json`, or set the an equivalent environment variable.
Because the flag is hidden, and there are no online references to the
functionality, I believe it is safe to remove. If users take a
dependency on JSON output, it should be enabled via the existing
`--output json` flag and be written to stdout.

After this change is merged, the remaining functionality can be moved
into the core `cmdio` types, and the "progress logger" can be removed.
Once there is a single type for all I/O, we have a better shot at
improving it.

Related change: #3811.

## Tests

Tests pass.