[Python] Move 'python' outside of experimental (#3540)
## Changes
Move `python` outside of the `experimental` section. 

The change is backward-compatible and forward-compatible. If CLI is
using a newer or older version of `databricks-bundles,` it will continue
to work. It's implemented by copying the content of `python` into
`experimental.python`. If we don't offer compatibility, there is a risk
that previously deployed resources will be erased because we wouldn't be
able to load resources.

See also https://github.com/databricks/cli/pull/3539

## Why
Python support for Databricks Asset Bundles is graduating from the
experimental status.

## Tests
Acceptance tests check that previous `databricks-bundles` versions
continue to work.