Include release date in CHANGELOG (#3846)
## Changes

Include release date in `CHANGELOG.md`.

The updates to the existing changelog were done with a JIT script that
compared the `git blame` output of the existing changelog with the tags
for those releases and added the corresponding date.

## Why

This makes it easier to eyeball when features or fixes were shipped.