Update lakeflow template project_name validation message (#3856)
## Changes
Updated to align with the pattern
## Why

Why not change the pattern to allow uppercase? It is lowercase because
this name is used for certain Python modules and functions. So we'd need
to accept that we're breaking the Python conventions if we allow upper
case

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->