Fix test creating `terraform.tfstate` in source directory (#3855)
## Changes

PR #3797 replaced a function that returns a state file path with one
that returns both a filename and a path. The test used only the
filename, causing the state file to be created in the wrong directory.

In the current state, it creates the file in `$PWD`, equal to the
directory where the test file is located. The path in `localPath` is
located inside the bundle cache directory, which is initialized to
somewhere inside `t.TempDir()` above.