Fix race condition in `cmdio.LogString` (#3868)
## Changes

Make `LogString` write the string and newline in a single call to reduce
the likelihood of interleaved output when called concurrently. This was
atomic until #3818 where I removed the lock, assuming it wasn't
necessary.

Note: This fix assumes `io.WriteString` provides atomic writes for small
messages, which is not guaranteed.

## Why

Fixes flaky test from:
https://github.com/databricks/cli/actions/runs/18995883239/job/54255615389