Add validation and translation for served_models and served_entities in serving endpoints (#3880)
## Changes && Why
`served_models` was deprecated and the fields in it were moved to
`served_endpoints`. The GET API response for a model serving endpoint
returns with both `served_models` and `served_entities` even if one is
set on the client side (not AIP complaint).

In order to simplify the direct deployment logic, this PR translates the
`served_models` config to `served_entities`, so that we can keep the
core diff logic in direct deployment simple.

Note: This is not a breaking change. TF also had the same logic with a
conflicts_with annotation for `served_models` and `served_entities`


## Tests
New acceptance tests. Manually confirmed that a deployment succeeds with
no issues.