Fix race condition in the ssh tunnel handover logic
Proxy client can't initiate multiple handovers at the same time from its point of view,
but proxy server can still be in the situation where 2 handovers racing for handoverMutex simultaneously.

The first one can still wait for the close-message ack from the client (on old ws connection), while the client
sends a new handover http request. Servers recieving loop has blockUntilHandoverComplete, which tries to grab a
handoverMutex, which races for it with the new acceptHandover call. If the later executed first, we have a deadlock.

The fix is better waitig logic on the reciever loop side - it cares to wait for a new connection to be activated,
and now it does exactly this using connChanged condition variable.