Fix race condition in the ssh tunnel handover logic (#3888)
## Changes
Proxy client can't initiate multiple handovers at the same time from its
point of view, but proxy server can still be in the situation where 2
handovers racing for handoverMutex simultaneously.

The first one can still wait for the close-message ack from the client
(on old ws connection), while the client sends a new handover http
request. Servers receiving loop has blockUntilHandoverComplete, which
tries to grab a handoverMutex, which races for it with the new
acceptHandover call. If the later executed first, we have a deadlock.

The fix is better waiting logic on the receiver loop side - it needs to
wait for a new connection to be activated, and now it does exactly this
using connSwapped channel.

Refactored the overall coordination logic a bit: we now use
handoverState atomic to store handover context (with a timeout), and two
channels necessary to coordinate between receiving loop and the handover
initiator. `isHandover` flag is not gone, as `handoverState` itself is
an indicator of when the handover is in progress.

### Testing
Also ran tests with `-race` and `-count 100` flags
\+ spawning manual ssh sessions with short handover timeout