Pass public key name in every client ssh connection (#3775)
## Changes
Before we were passing a client key name (a secret name that contains
the actual client public key) to the ssh server job. After it's started,
the ssh server was only accepting one client with one specific key. This
is a problem when you try to ssh to the same cluster under the same
databricks user from multiple different client machines (that have
different ssh keys).

Our integration tests can't work consistently because of this, and we've
already got one customer report about this problem.

The main change here is that now we don't always generate key-pairs on
the client, but check if the pair already exists in the secret scope and
use it. And if not, we generate the pair and save both keys to the scope
(and before we were only saving public key there)