Add default scala template to databricks cli (#3906)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
This adds a new default scala DABs template as a follow up to
https://github.com/databricks/bundle-examples/pull/119
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This is to provide an off the shelf template for customers to start
using scala dbconnect to develop scala jobs in Databricks

## Tests
<!-- How have you tested the changes? -->
Manually tested user flow interactively and as job workload for both
standard and serverless
<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->