Add ready channel to proxyConnection for initial connection synchronization
This change introduces a new channel in the proxyConnection struct to signal when the initial connection is established or has failed. It prevents race conditions during handover by ensuring that handover is only accepted after the initial connection is ready. This is not really a reall world scenario, but good to handle just for the sake of tests