Show permissions & grants in plan; use resource keys in the output (#3961)
## Changes
The json plan for terraform now includes grants and permissions
resources and skip actions. This matches direct plan.

The text plan for terraform and direct now print permissions and grants
as separate resources.

The "The following resources will be deleted" message now mention
resources by full name, e.g. "resources.sql_warehouses.foo" instead of
"sql_warehouse foo". This matches "bundle plan" output and internal
representation. Note, this message still filters out permissions and
grants, they are not relevant there.

For context, we used to use terraform-based resource types and shows
grants & permissions in "will be deleted" message:
https://github.com/databricks/cli/pull/3075

## Why
Consistency: we already use "resources.jobs.foo" format in json plan,
direct state file, error messages. This is extends it to the rest of the
output.

We need to treat grants and permissions as first class citizens, because
they need to be migrated. After migration we're going to ask users to
run "bundle plan" to check for permanent drift and it's important that
we show changes in permissions/grants there and also that we separate
them from their parent resource.


## Tests
Existing tests, some are extended to run additional "bundle plan"
commands.