Fix a potential panic during `databricks labs install --offline=true` (#3955)
## Changes

This PR fixes a panic during `databricks labs install --offline=true` if
the release cache can't be loaded. (Preliminary support for offline
installs was introduced by #2049.)

## Why

This is easy to hit during offline installation if the archive hasn't
been unpacked into the location expected by the code. When this happens
the code should not panic but rather propagate the error.

## Tests

Tested manually, and via some additional tests that cover loading the
release cache in offline mode:

 - A test for the happy path (already covered indirectly)
- A new test for the error handling (that would have revealed the
panic).

Co-authored-by: Pieter Noordhuis <pieter.noordhuis@databricks.com>