MCP: Streamline validation for node.js templates (#3985)
## Changes

Validation for 'overfit' for the tRPC template. I generalized it to
follow node.js standards that all node.js templates can implement.

Before: 
* Build
  * npm run build
* Typecheck
  * cd client && npx tsc --noEmit
* Test
  * npm test

After:
* Build
  * npm run build –if-present
* Typecheck
  * npm run typecheck –if-present
* Test
  * npm run test –if-present



## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->