feat(apps): Add apps new dev server (#3923)
## Changes
<!-- Brief summary of your changes that is easy to understand -->

This PR introduces a new way to develop apps, that fully integrates with
the new Apps Kit. It works along with this other PR
https://github.com/databricks-eng/apps-sdk/pull/23 which is already
merged.

The idea is to provide an easy way of developing applications without
having to setup all the environment variables and all the dependencies.
This creates an hybrid between running the local frontend code + the
remote server.

It also provides integration with HMR, so when any change is make either
in the frontend code or query files, they will get reflected instantly
on the app. (Changes are not synced, but the user can see them)

### How does it work?

You can find a diagram of how this works in the other
[PR](https://github.com/databricks-eng/apps-sdk/pull/23)

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

This is a step towards making the experience of developing Databricks
apps a lot easier for users.

Basically with the CLI + the Apps Kit, having a first deployment of the
app into the Databricks environment, by simply running the new
`dev-remote` command in the project folder, now users can easily preview
their changes against the remote environment without setting anything
up.

## Tests
<!-- How have you tested the changes? -->
- Added some tests to verify some of the functionality, path parsing,
etc.
- Did a lot of manual testing to verify everything works as expected

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->

---------

Co-authored-by: MarioCadenas <MarioCadenas@users.noreply.github.com>